Move stack checking to rewriters
