+++ src/Compilers/VisualBasic/Portable/Syntax/VisualBasicSyntaxRewriter.vb
@@ -29,6 +29,26 @@ Namespace Microsoft.CodeAnalysis.VisualBasic
             End Get
         End Property
 
+        Private Const MaxUncheckedRecursionDepth As Integer = Syntax.InternalSyntax.Parser.MaxUncheckedRecursionDepth
+        Private _recursionDepth As Integer
+
+        Public Overrides Function Visit(node As SyntaxNode) As SyntaxNode
+            If node IsNot Nothing Then
+                _recursionDepth += 1
+
+                If _recursionDepth > MaxUncheckedRecursionDepth Then
+                    PortableShim.RuntimeHelpers.EnsureSufficientExecutionStack()
+                End If
+
+                Dim result = DirectCast(node, VisualBasicSyntaxNode).Accept(Me)
+
+                _recursionDepth -= 1
+                Return result
+            Else
+                Return node
+            End If
+        End Function
+
         Public Overridable Function VisitToken(token As SyntaxToken) As SyntaxToken
             Dim leading = Me.VisitList(token.LeadingTrivia)
             Dim trailing = Me.VisitList(token.TrailingTrivia)
