+++ src/Compilers/VisualBasic/Portable/Syntax/VisualBasicSyntaxVisitor.vb
@@ -10,20 +10,10 @@ Namespace Microsoft.CodeAnalysis.VisualBasic
     ''' passed into its <see cref="Visit(SyntaxNode)"/> method.
     ''' </summary>
     Partial Public MustInherit Class VisualBasicSyntaxVisitor
-        Private Const MaxUncheckedRecursionDepth As Integer = Syntax.InternalSyntax.Parser.MaxUncheckedRecursionDepth
-        Private _recursionDepth As Integer
 
         Public Overridable Sub Visit(ByVal node As SyntaxNode)
             If node IsNot Nothing Then
-                _recursionDepth += 1
-
-                If _recursionDepth > MaxUncheckedRecursionDepth Then
-                    PortableShim.RuntimeHelpers.EnsureSufficientExecutionStack()
-                End If
-
                 DirectCast(node, VisualBasicSyntaxNode).Accept(Me)
-
-                _recursionDepth -= 1
             End If
         End Sub
 
@@ -40,22 +30,10 @@ Namespace Microsoft.CodeAnalysis.VisualBasic
     ''' The type of the return value of this visitor's Visit method.
     ''' </typeparam>
     Partial Public MustInherit Class VisualBasicSyntaxVisitor(Of TResult)
-        Private Const MaxUncheckedRecursionDepth As Integer = Syntax.InternalSyntax.Parser.MaxUncheckedRecursionDepth
-        Private _recursionDepth As Integer
 
         Public Overridable Function Visit(ByVal node As SyntaxNode) As TResult
             If node IsNot Nothing Then
-                _recursionDepth += 1
-
-                If _recursionDepth > MaxUncheckedRecursionDepth Then
-                    PortableShim.RuntimeHelpers.EnsureSufficientExecutionStack()
-                End If
-
-                Dim result = DirectCast(node, VisualBasicSyntaxNode).Accept(Me)
-
-                _recursionDepth -= 1
-
-                Return result
+                Return DirectCast(node, VisualBasicSyntaxNode).Accept(Me)
             End If
 
             Return Nothing
